+++ aerich/ddl/sqlite/__init__.py
@@ -1,4 +1,4 @@
-from typing import Type
+from __future__ import annotations
 
 from tortoise import Model
 from tortoise.backends.sqlite.schema_generator import SqliteSchemaGenerator
@@ -17,14 +17,14 @@ class SqliteDDL(BaseDDL):
     )
     _DROP_UNIQUE_TEMPLATE = 'DROP INDEX IF EXISTS "{column_name}_2"'
 
-    def modify_column(self, model: "Type[Model]", field_object: dict, is_pk: bool = True):
+    def modify_column(self, model: type[Model], field_object: dict, is_pk: bool = True):
         raise NotSupportError("Modify column is unsupported in SQLite.")
 
-    def alter_column_default(self, model: "Type[Model]", field_describe: dict):
+    def alter_column_default(self, model: type[Model], field_describe: dict):
         raise NotSupportError("Alter column default is unsupported in SQLite.")
 
-    def alter_column_null(self, model: "Type[Model]", field_describe: dict):
+    def alter_column_null(self, model: type[Model], field_describe: dict):
         raise NotSupportError("Alter column null is unsupported in SQLite.")
 
-    def set_comment(self, model: "Type[Model]", field_describe: dict):
+    def set_comment(self, model: type[Model], field_describe: dict):
         raise NotSupportError("Alter column comment is unsupported in SQLite.")